service: Improve description of natural order and ranking order
We improve the javadoc for ServiceReference.compareTo for the natural
order and state that the ranking order, which is defined in the Core
spec, is the reverse of the natural order.

We then update the javadoc and spec text to consistently use the terms
natural order or ranking order, as appropriate, instead of restating
how to sort on service.ranking and service.id service properties.

Fixes https://github.com/osgi/osgi/issues/400

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>